Move documentation tests
